[#276] Refactor clangd API
- internal package names according to eclipse platform naming convention

part of #276
