+++ bundles/org.eclipse.cdt.lsp.clangd/META-INF/MANIFEST.MF
@@ -3,12 +3,8 @@ Automatic-Module-Name: org.eclipse.cdt.lsp.clangd
 Bundle-ActivationPolicy: lazy
 Bundle-ManifestVersion: 2
 Bundle-SymbolicName: org.eclipse.cdt.lsp.clangd;singleton:=true
-Bundle-Version: 1.0.0.qualifier
-Export-Package: org.eclipse.cdt.lsp.clangd;x-internal:=true,
- org.eclipse.cdt.lsp.internal.clangd;x-internal:=true,
- org.eclipse.cdt.lsp.internal.clangd.editor;x-internal:=true,
- org.eclipse.cdt.lsp.internal.clangd.editor.expressions;x-internal:=true,
- org.eclipse.cdt.lsp.internal.clangd.editor.handlers;x-internal:=true
+Bundle-Version: 2.0.0.qualifier
+Export-Package: org.eclipse.cdt.lsp.clangd
 Import-Package: org.yaml.snakeyaml;version="1.27.0",
  org.yaml.snakeyaml.error;version="1.27.0",
  org.yaml.snakeyaml.scanner;version="1.27.0"
@@ -32,9 +28,9 @@ Require-Bundle: org.eclipse.cdt.lsp;bundle-version="0.0.0",
  org.eclipse.ui.workbench.texteditor;bundle-version="0.0.0",
  org.eclipse.core.variables;bundle-version="0.0.0",
  org.yaml.snakeyaml;bundle-version="0.0.0"
-Service-Component: OSGI-INF/org.eclipse.cdt.lsp.clangd.BuiltinClangdOptionsDefaults.xml,
- OSGI-INF/org.eclipse.cdt.lsp.clangd.ClangdConfigurationFileManager.xml,
- OSGI-INF/org.eclipse.cdt.lsp.internal.clangd.ClangdConfigurationAccess.xml,
- OSGI-INF/org.eclipse.cdt.lsp.internal.clangd.ClangdFallbackManager.xml,
- OSGI-INF/org.eclipse.cdt.lsp.internal.clangd.ClangdMetadataDefaults.xml
-Bundle-Activator: org.eclipse.cdt.lsp.internal.clangd.editor.ClangdPlugin
+Service-Component: OSGI-INF/org.eclipse.cdt.lsp.clangd.internal.config.BuiltinClangdOptionsDefaults.xml,
+ OSGI-INF/org.eclipse.cdt.lsp.clangd.internal.config.ClangdConfigurationAccess.xml,
+ OSGI-INF/org.eclipse.cdt.lsp.clangd.internal.config.ClangdConfigurationFileManager.xml,
+ OSGI-INF/org.eclipse.cdt.lsp.clangd.internal.config.ClangdFallbackManager.xml,
+ OSGI-INF/org.eclipse.cdt.lsp.clangd.internal.config.ClangdMetadataDefaults.xml
+Bundle-Activator: org.eclipse.cdt.lsp.clangd.plugin.ClangdPlugin
