+++ bundles/org.eclipse.cdt.lsp.clangd/OSGI-INF/org.eclipse.cdt.lsp.clangd.internal.config.ClangdFallbackManager.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.3.0" name="org.eclipse.cdt.lsp.internal.clangd.ClangdFallbackManager">
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.3.0" name="org.eclipse.cdt.lsp.clangd.internal.config.ClangdFallbackManager">
    <service>
       <provide interface="org.eclipse.cdt.lsp.clangd.ClangdFallbackFlags"/>
    </service>
    <reference cardinality="1..1" field="build" interface="org.eclipse.cdt.core.build.ICBuildConfigurationManager" name="build"/>
    <reference cardinality="1..1" field="uri" interface="org.eclipse.cdt.lsp.editor.InitialUri" name="uri"/>
    <reference cardinality="1..1" field="workspace" interface="org.eclipse.core.resources.IWorkspace" name="workspace"/>
-   <implementation class="org.eclipse.cdt.lsp.internal.clangd.ClangdFallbackManager"/>
+   <implementation class="org.eclipse.cdt.lsp.clangd.internal.config.ClangdFallbackManager"/>
 </scr:component>
