+++ bundles/org.eclipse.cdt.lsp.clangd/src/org/eclipse/cdt/lsp/clangd/internal/config/ClangdMetadataDefaults.java
@@ -11,13 +11,13 @@
  * Contributors:
  *     Alexander Fedorov (ArSysOp) - Initial API and implementation
  *******************************************************************************/
-package org.eclipse.cdt.lsp.internal.clangd;
+package org.eclipse.cdt.lsp.clangd.internal.config;
 
 import java.util.stream.Collectors;
 
 import org.eclipse.cdt.lsp.clangd.ClangdMetadata;
 import org.eclipse.cdt.lsp.clangd.ClangdOptionsDefaults;
-import org.eclipse.cdt.lsp.internal.clangd.editor.LspEditorUiMessages;
+import org.eclipse.cdt.lsp.clangd.internal.ui.LspEditorUiMessages;
 import org.eclipse.core.runtime.preferences.PreferenceMetadata;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
